refactor(credit): generic transaction
