+++ internal/credit/postgres_connector/watermark.go
@@ -23,19 +23,6 @@ func (c *PostgresConnector) GetHighWatermark(ctx context.Context, namespace stri
 
 	if err != nil {
 		if db.IsNotFound(err) {
-			// Upsert ledger for the subject
-			err := c.db.Ledger.
-				Create().
-				SetNamespace(namespace).
-				SetSubject(subject).
-				SetHighwatermark(defaultHighwatermark).
-				OnConflictColumns(db_ledger.FieldNamespace, db_ledger.FieldSubject).
-				Ignore().
-				Exec(ctx)
-			if err != nil {
-				return credit_model.HighWatermark{}, fmt.Errorf("failed to upsert ledger: %w", err)
-			}
-
 			return credit_model.HighWatermark{
 				Subject: subject,
 				Time:    defaultHighwatermark,
@@ -51,16 +38,10 @@ func (c *PostgresConnector) GetHighWatermark(ctx context.Context, namespace stri
 	}, nil
 }
 
-// GetHighWatermark returns the high watermark for the given credit and subject pair.
-func (c *PostgresConnector) checkHighWatermark(ctx context.Context, namespace string, subject string, t time.Time) (credit_model.HighWatermark, error) {
-	hw, err := c.GetHighWatermark(ctx, namespace, subject)
-	if err != nil {
-		return credit_model.HighWatermark{}, err
+func checkAfterHighWatermark(t time.Time, ledger *db.Ledger) error {
+	if !t.After(ledger.Highwatermark) {
+		return &credit_model.HighWatermarBeforeError{Namespace: ledger.Namespace, Subject: ledger.Subject, HighWatermark: ledger.Highwatermark}
 	}
 
-	if !t.After(hw.Time) {
-		return credit_model.HighWatermark{}, &credit_model.HighWatermarBeforeError{Namespace: namespace, Subject: subject, HighWatermark: hw.Time}
-	}
-
-	return hw, nil
+	return nil
 }
